requirements.txt
selenium>=4.15.0
requests>=2.31.0
aiohttp>=3.9.0
pyodbc>=5.0.0
pyyaml>=6.0
webdriver-manager>=4.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
